% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cv_descriptives.R
\name{bmean}
\alias{bmean}
\title{Wrapper para utilizar la función \code{mean()} como base para calcular intervalos de confienza mediante bootstrap}
\usage{
bmean(data, indices)
}
\arguments{
\item{data}{Un dataframe con los datos de la encuesta}

\item{indices}{argument interno para realizar el bootstrap}
}
\value{
un \code{numeric} con la media
}
\description{
Wrapper para utilizar la función \code{mean()} como base para calcular intervalos de confienza mediante bootstrap
}
