% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cv_descriptives.R
\name{bmedian}
\alias{bmedian}
\title{Wrapper para utilizar la función \code{median()} como base para calcular intervalos de confienza mediante bootstrap}
\usage{
bmedian(data, indices)
}
\arguments{
\item{data}{Un dataframe con los datos de la encuesta}

\item{indices}{argument interno para realizar el bootstrap}
}
\value{
un \code{vector} con los cuantiles especificados
}
\description{
Wrapper para utilizar la función \code{median()} como base para calcular intervalos de confienza mediante bootstrap
}
