% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cv_descriptives.R
\name{boot_descriptive}
\alias{boot_descriptive}
\title{Realiza la estimación del coeficiente de variación y los intervalos de confianza sobre una medida descriptiva
-por ahora, \code{bmean}, \code{bquantile}, \code{sd}-.}
\usage{
boot_descriptive(data, col, statistic, ci_alpha = 0.95, R = 1000,
  strata = NULL)
}
\arguments{
\item{data}{Un dataframe con los datos de la encuesta}

\item{col}{variable sobre la que se busca calcular la media, Debe pasarse entre comillas}

\item{statistic}{estadísticos obre el que se quiere calcular el cv y cint}

\item{ci_alpha}{numeric. alpha para el cálculo del intervalo de confianza}

\item{R}{cantidad de repeticiones bootstrap}

\item{strata}{variable con la identificación de los estratos}
}
\value{
una lista
}
\description{
Realiza la estimación del coeficiente de variación y los intervalos de confianza sobre una medida descriptiva
-por ahora, \code{bmean}, \code{bquantile}, \code{sd}-.
}
