% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cv_descriptives.R
\name{bquantile}
\alias{bquantile}
\title{NO FUNCIONAL Wrapper para utilizar la función \code{quantile()} como base para calcular intervalos de confienza mediante bootstrap}
\usage{
bquantile(data, q = c(0.25, 0.5, 0.75), indices)
}
\arguments{
\item{data}{Un dataframe con los datos de la encuesta}

\item{q}{vector con las probabilidades de los cuantiles a calcular.}

\item{indices}{argument interno para realizar el bootstrap}
}
\value{
un \code{vector} con los cuantiles especificados
}
\description{
NO FUNCIONAL Wrapper para utilizar la función \code{quantile()} como base para calcular intervalos de confienza mediante bootstrap
}
