% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cv_table.R
\name{calc_cvs_table}
\alias{calc_cvs_table}
\title{Calcula coef de variación de una tabla de contingencia en una lista.}
\usage{
calc_cvs_table(r, table_orig)
}
\arguments{
\item{r}{resultado de la aplicación del bootstrap}

\item{table_orig}{tabla original a boostrapear}
}
\value{
una lista
}
\description{
Calcula coef de variación de una tabla de contingencia en una lista.
}
