% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cv_table.R
\name{cv_raw_freq}
\alias{cv_raw_freq}
\title{Calcula coeficientes de variación para una distribución de frecuencias univariada, mediante bootstrap}
\usage{
cv_raw_freq(data, R, strata, var)
}
\arguments{
\item{data}{Una tabla o dataframe con los datos de la encuesta}

\item{R}{cantidad de repeticiones boostrap}

\item{strata}{variable con el identificador de estratos}

\item{var}{string con las variables que conforman la tabla de contingencia. Deben estar especficiados en formato de fórmula: \code{y~x}}
}
\value{
una lista con dos elementos: la tabla original y la tabla de coeficientes de variación
}
\description{
Calcula coeficientes de variación para una distribución de frecuencias univariada, mediante bootstrap
}
