% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_table.R
\name{cv_raw_table}
\alias{cv_raw_table}
\title{Calcula coeficientes de variaci贸n para una tabla de contingencia, mediante bootstrap}
\usage{
cv_raw_table(data, R, strata, formula)
}
\arguments{
\item{data}{Una tabla o dataframe con los datos de la encuesta}

\item{R}{cantidad de repeticiones boostrap}

\item{strata}{variable con el identificador de estratos}

\item{formula}{variables que conforman la tabla de contingencia. Deben estar especficiados en formato de f贸rmula: \code{y~x}}
}
\value{
una lista con dos elementos: la tabla original y la tabla de coeficientes de variaci贸n
}
\description{
Calcula coeficientes de variaci贸n para una tabla de contingencia, mediante bootstrap
}
