% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dataset.R
\name{format_dataset}
\alias{format_dataset}
\title{Realiza el formateo de la base de datos, generando la variable de identificación del estrato.}
\usage{
format_dataset(path)
}
\arguments{
\item{path}{string que contiene la ruta al archivo que contiene la base.}
}
\value{
un dataframe formateado
}
\description{
Realiza el formateo de la base de datos, generando la variable de identificación del estrato.
}
