% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cv_table.R
\name{format_results}
\alias{format_results}
\title{Formatea los resultados en una lista.}
\usage{
format_results(r, table_orig)
}
\arguments{
\item{r}{resultado de la aplicación del bootstrap}

\item{table_orig}{tabla original a boostrapear}
}
\value{
una lista
}
\description{
Formatea los resultados en una lista.
}
