% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cv_table.R
\name{format_results_table}
\alias{format_results_table}
\title{Formatea resultados para la salida final}
\usage{
format_results_table(table_orig, r, cvs, cv_final)
}
\arguments{
\item{table_orig}{tabla original a boostrapear}

\item{r}{resultado de la aplicación del bootstrap}

\item{cvs}{cvs calculados}

\item{cv_final}{matriz o vector con calculos de cv}
}
\value{
una lista
}
\description{
Formatea resultados para la salida final
}
