% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cv_table.R
\name{gen_frequence}
\alias{gen_frequence}
\title{Genera una distribución de frecuencias univariada}
\usage{
gen_frequence(var, data, indices)
}
\arguments{
\item{var}{string con las variables que conforman la tabla de contingencia. Deben estar especficiados en formato de fórmula: \code{y~x}}

\item{data}{Una tabla o dataframe con los datos de la encuesta}

\item{indices}{argument interno para realizar el bootstrap}
}
\value{
una tabla original
}
\description{
Genera una distribución de frecuencias univariada
}
