% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_descriptives.R
\name{bquantile}
\alias{bquantile}
\title{Wrapper para utilizar la función \code{quantile()} como base para calcular intervalos de confienza mediante bootstrap}
\usage{
bquantile(data, col, q, indices)
}
\arguments{
\item{data}{Un dataframe con los datos de la encuesta}

\item{col}{variable sobre la que se busca calcular la media, Debe pasarse entre comillas}

\item{q}{número de cuantiles a calcular. q=}

\item{indices}{argument interno para realizar el bootstrap}
}
\value{
un \code{vector} con los cuantiles especificados
}
\description{
Wrapper para utilizar la función \code{quantile()} como base para calcular intervalos de confienza mediante bootstrap
}
