% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_descriptives.R
\name{bsd}
\alias{bsd}
\title{Wrapper para utilizar la función \code{sd()} como base para calcular intervalos de confienza mediante bootstrap}
\usage{
bsd(data, col, indices)
}
\arguments{
\item{data}{Un dataframe con los datos de la encuesta}

\item{col}{variable sobre la que se busca calcular la media, Debe pasarse entre comillas}

\item{indices}{argument interno para realizar el bootstrap}
}
\value{
un \code{numeric} con el desvío estándar
}
\description{
Wrapper para utilizar la función \code{sd()} como base para calcular intervalos de confienza mediante bootstrap
}
